Refactor totalCount to cartItems
totalCount would allow us to properly fulfill the base requirements of
the project description, however it would not accurately reflect a
real-world scenario. Thus it was refactored to a cartItems array and
the length of the array is used instead. This opens up the possibility
to add a ShoppingCart page in the future (as opposed to a component
which only displays a number in the NavBar component).