Fix bug that increased shopping cart to 100
If you typed in the input, for example, "4", "1", "1", the expect result
would be for the number to remain at "41" as a cap of 100 was
arbitrarily added to each item. The input value itself would remain at
"41" however the shopping cart number at the top of the screen would
jump to 100.